LargestProductInSeries Complete.
There is an optimization we can do if the product is 0.
We are following a sliding window method (0-13, 1-14, ... ) if any of
the number lists' product is 0, we can pivot the 13 numebr window like
fix the window position where it is 0 and rotate it to next 13 elements
starting from that index. Think about it.